+++ changelog.md
@@ -4,10 +4,12 @@
 
 ### Added
 - One new hooks/actions added in layouts: `lsx_entry_inside_top` to hook into a page content.
+- WooCommerce 4.1 compatibility.
+- Adding Yoast 14.0.+ compatibility.
 
 ### Changed
-- Woocommerce checkout and cart styling improved.
-- Layout changed on Checkout and cart woocommerce pages.
+- Woocommerce Cart, Checkout, Thank You and Restricted styling improved.
+- Layout changed on Cart, Checkout, Thank You and Restricted woocommerce pages.
 
 ### Fixed
 - Fixed issue #346 'If LSX banners is not enabled there is a simple banner showing on each page.'.