+++ changelog.md
@@ -1,6 +1,6 @@
 # LSX Change log
 
-# [[2.7.1]] - In Development
+# [[2.8.0]] - 2020-05-21
 
 ### Added
 - One new hooks/actions added in layouts: `lsx_entry_inside_top` to hook into a page content.
@@ -17,6 +17,9 @@
 - Fixed issue #346 'If LSX banners is not enabled there is a simple banner showing on each page.'.
 - Body class `using-gutenberg` affecting other custom post type single pages was removed from any custom post type single page.
 - Fixed a few styling issues related to Sensei templates.
+
+### Security
+- Testing compatibility with WordPress 5.4.1
 
 
 ## [[2.7.0]](https://github.com/lightspeeddevelopment/lsx/releases/tag/2.7.0) - 2020-03-30